[SYSTEMDS-2624] Rework acquireRead and cleanup of federated matrices
The integration of federated matrices and frames into our caching and
buffer pool framework was suboptimal because (1) repeated pinning of a
federated data into memory of the coordinator caused repeated
consolidation of the federated data, and (2) missing cleanup of
federated data intermediates. This rework fixes both issues and thus,
reduces unnecessary memory pressure at the workers and unnecessary data
transfer at the coordinator.

For the cleanup of federated data, we added special guards to ensure
only federated intermediates (and files thereof) are deleted but NOT the
original federated input files.

+++ src/main/java/org/apache/sysds/runtime/controlprogram/context/ExecutionContext.java
@@ -59,7 +59,9 @@ import org.apache.sysds.utils.Statistics;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 public class ExecutionContext {
@@ -71,6 +73,7 @@ public class ExecutionContext {
 	//symbol table
 	protected LocalVariableMap _variables;
 	protected boolean _autoCreateVars;
+	protected Set<String> _guardedFiles = new HashSet<>();
 	
 	//lineage map, cache, prepared dedup blocks
 	protected Lineage _lineage;
@@ -130,6 +133,10 @@ public class ExecutionContext {
 	
 	public void setAutoCreateVars(boolean flag) {
 		_autoCreateVars = flag;
+	}
+	
+	public void addGuardedFilename(String fname) {
+		_guardedFiles.add(fname);
 	}
 
 	/**
@@ -751,7 +758,7 @@ public class ExecutionContext {
 			//compute ref count only if matrix cleanup actually necessary
 			if ( mo.isCleanupEnabled() && !getVariables().hasReferences(mo) )  {
 				mo.clearData(); //clean cached data
-				if( fileExists ) {
+				if( fileExists && !_guardedFiles.contains(mo.getFileName()) ) {
 					HDFSTool.deleteFileIfExistOnHDFS(mo.getFileName());
 					HDFSTool.deleteFileIfExistOnHDFS(mo.getFileName()+".mtd");
 				}
